backend/requirements.txt
# Core web framework
flask==3.0.0
flask-cors==4.0.0

# Machine Learning & RL
torch==2.1.0
stable-baselines3==2.2.1
gymnasium==0.29.1
numpy==1.24.4
pandas==2.1.1

# Trading & Market Data
alpaca-py==0.21.0
yfinance==0.2.28

# Technical Analysis
TA-Lib==0.4.28
ta==0.10.2

# NLP & Sentiment Analysis
transformers==4.35.2
torch-audio==2.1.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pusher==3.3.2
celery==5.3.4
redis==5.0.1

# Development
pytest==7.4.3
black==23.9.1
flake8==6.1.0